refactor(permissions): move lookup helpers beside their hooks
get_permission and get_permission_scope were plain in-memory lookups living among
the async DB functions in utils.py. Move each next to the hook it reads:
get_permission into __init__.py by PERMISSIONS, get_permission_scope into scopes.py
by PERMISSION_SCOPES. PermissionScope.create() now reuses get_permission_scope for
its parent check instead of inlining the same lookup (which it did only to avoid a
utils import cycle). The lib facade re-exports both, so the public API is unchanged.

Also hoist the EMAIL_WHITELIST_READ import to the top of test_service.py (it had no
reason to be a local import) and repoint the helper tests' monkeypatch targets to the
hooks' new module namespaces.
