+++ app/lib/permissions/__init__.py
@@ -17,13 +17,13 @@ from app.core.permissions import (
     EMAIL_WHITELIST_DELETE,
     EMAIL_WHITELIST_READ,
     Permission,
+    get_permission,
 )
 from app.core.permissions.exceptions import PermissionNotFound, PermissionScopeNotFound, RoleNotFound
+from app.core.permissions.scopes import get_permission_scope
 from app.core.permissions.utils import (
     assign_role,
     can,
-    get_permission,
-    get_permission_scope,
     has_role,
     revoke_role,
 )
